(BE) [fix] api 경로 수정
